Move wilight base entity to separate module
