+++ homeassistant/components/wilight/const.py
@@ -1,0 +1,3 @@
+"""The WiLight integration."""
+
+DOMAIN = "wilight"
